test: move /users tests to separate file
